[sftp] Rename obsolete file names
